// CabBookingSPA/src/app/shared/models/bookingModel.ts
// Generated by https://quicktype.io
import { Location } from "./locationModel";
export interface Booking {
    id:          number;
    email:       string;
    bookingDate: string;
    bookingTime: string;
    pickupDate:  string;
    pickupTime:  string;
    contactNo:   string;
    status:      string;
    cabTypesId:  number;
    location:   Location;
}